Move the tangelo parsing file into the io directory
